Add quotes to .env.dist and clear ENCRYPTION_KEY
All build steps currently fail as the .env.dist is invalid,
due to values with whitespace in them.

In CI, the .env.dist is copied to .env and used from there. The provided
`[Paste from 1Password]` entry is an invalid hexadecimal string, and
PHPStan crashes when it tries to parse it when evaluating the
EncryptionService.