Use CastorServer's Client Credentials when available
When authenticating with an EDC server, we now use the Client Credentials
available in the database for the given server. This allows us to use
different client credentials per server, so that we don't need to copy
over client credentials from EDC database to EDC database.

We do this by overriding the AbstractProvider::clientId/clientSecret
when we start the authorization code and access token requests in the
CastorUserProvider.

If there are no client credentials available for a given server, we still
fallback to the ones specified in the .env file.