ALlow authenticating via EDC access tokens
This commit adds a new Authenticator the FDP, that allows users to run API requests with a valid EDC API token. The token will be validated against the server for which it was issued, and if valid, is allowed to continue accessing the API.

It requires requests to provide two new headers:

    X-AUTH-TOKEN: a valid access token for an EDC installation
    X-AUTH-SERVER: the full host (including protocol) of the server for which the access token is valid.

Under the hood, at authentication time, the Authenticator will issue an /api/user request using the provided X-AUTH-TOKEN. If valid, that will return the user information, if not we get back an error and issue a 401 Forbidden to the end user.

To ensure the FDP already knows about the user, we check for existence of the UserID in the CastorUser repository. If it doesn't exist yet, we return a 401 with an error message as well. Reason for this is that everywhere through the application we require the profile to be completed via the UserWizardSubscriber, and all API calls will be forwarded to the user wizard.